+++ src/app/component-library/property-grid/property-grid/property-grid-field.ts
@@ -21,4 +21,5 @@ export interface PropertyGridField {
   additionalClasses?: string;
 
   copyButton?: boolean;
+  tooltip?: string | null;
 }