+++ src/app/core/config/profiles/feature-sets/enterprise-edition-features.ts
@@ -2,6 +2,6 @@ import {EdcUiFeature} from '../edc-ui-feature';
 
 export const ENTERPRISE_EDITION_FEATURES: EdcUiFeature[] = [
   'logout-button',
-  'connector-limit-for-consuming-contract-agreements',
+  'connector-limits',
   'sovity-zammad-integration',
 ];