+++ src/app/routes/connector-ui/contract-agreement-page/contract-agreement-cards/contract-agreement-card-mapped.service.ts
@@ -27,8 +27,8 @@ export class ContractAgreementCardMappedService {
       isInProgress: contractAgreement.transferProcesses.some(
         (it) => it.state.simplifiedState === 'RUNNING',
       ),
-      isActivated: true,
-      connectorMaxLimit: null,
+      statusText: '',
+      statusTooltipText: '',
       searchTargets: [
         contractAgreement.contractAgreementId,
         contractAgreement.counterPartyId,