+++ src/app/routes/connector-ui/contract-agreement-page/contract-agreement-cards/contract-agreement-card-mapped.ts
@@ -7,7 +7,7 @@ export type ContractAgreementCardMapped = Omit<
 > & {
   asset: Asset;
   isInProgress: boolean;
-  isActivated: boolean;
-  connectorMaxLimit: number | null;
+  statusText: string;
+  statusTooltipText: string | null;
   searchTargets: (string | null)[];
 };